+++ app_estudo/integrations/__init__.py
@@ -3,6 +3,7 @@
 from .ankiconnect_healthcheck import AnkiHealthcheckResult, check_ankiconnect
 from .anki_mapping import LogicalAnkiNote, map_to_anki_logical_note
 from .ankiconnect_client import AnkiConnectClient, AnkiSyncResult
+from .anki_reconciliation import DuplicateGroup, ReconciliationReport, build_duplicate_groups
 
 __all__ = [
 	"LogicalAnkiNote",
@@ -11,4 +12,7 @@ __all__ = [
 	"check_ankiconnect",
 	"AnkiConnectClient",
 	"AnkiSyncResult",
+	"DuplicateGroup",
+	"ReconciliationReport",
+	"build_duplicate_groups",
 ]